+++ src/app/components/category-picker/terra-category-picker.component.ts
@@ -40,6 +40,12 @@ export class TerraCategoryPickerComponent implements OnInit, AfterContentChecked
     @Input()
     public inputIsDisabled:boolean;
 
+    /**
+     * @description Tooltip that is shown on the TextInput
+     */
+    @Input()
+    public inputTooltipText:string;
+
     @Input()
     public inputName:string;
 
@@ -72,7 +78,7 @@ export class TerraCategoryPickerComponent implements OnInit, AfterContentChecked
             this.categoryTreeConfig.list = this._list;
         }
 
-        if(!isNullOrUndefined(this.categoryTreeConfig.currentSelectedNode))
+        if(!isNullOrUndefined(this.categoryTreeConfig.currentSelectedNode) && !isNullOrUndefined(this.categoryTreeConfig.currentSelectedNode.value))
         {
             this._isContainerCategorySelected = (this.categoryTreeConfig.currentSelectedNode.value.type === 'container');
         }