+++ CHANGELOG.md
@@ -5,6 +5,33 @@
 * **terra-text-input**
 	- changed color of text input with readonly attribute
 	- rearranged style selectors to match native html attributes
+* **category-picker** fixed error with the reset button
+
+<a name="2.2.5"></a>
+# 2.2.5 (04.05.2018)
+
+### Feature
+* **terra-multi-split-view** set input inputComponentRoute deprecated. It is no longer needed.
+
+### Bug Fixes
+* **terra-converter-helper** updated a used pipe after the angular update
+
+<a name="2.2.4"></a>
+# 2.2.4 (03.05.2018)
+
+### Feature
+* **terra-select-box** fixed null pointer when <code>selectedValue</code> was undefined
+
+<a name="2.2.3"></a>
+# 2.2.3 (02.05.2018)
+
+### Feature
+* **string-helper** new helper for strings
+
+### Bug Fixes
+* **terra-text-input** changed color of text input with readonly attribute
+* **terra-dynamic-form** added tooltip for category picker
+* **terra-category-picker** added input for tooltip
 
 <a name="2.2.2"></a>
 # 2.2.2 (25.04.2018)
@@ -16,7 +43,7 @@
 # 2.2.1 (24.04.2018)
 
 ### Bug Fixes
-* **themes** theming restructuring
+* **tehmes** theming restructuring
 
 ## Feature
 * **terra-file-chooser** Added new component to open file-browser directly via a button without input element.